VError - refactor validation
