+++ frontend-vue/src/components/common/BaseAddEditDialog.vue
@@ -1,8 +1,12 @@
 <script>
 import { useVuelidate } from '@vuelidate/core';
 
+import VError from '@/components/common/VError';
+
 export default {
     name: 'BaseAddEditDialog',
+
+    extends: VError,
 
     props: {
         isOpened: {
@@ -27,8 +31,7 @@ export default {
         return {
             defaultForm,
             formData: { ...defaultForm },
-            isDialogOpened: false,
-            serverErrors: []
+            isDialogOpened: false
         };
     },
 
@@ -50,11 +53,6 @@ export default {
     },
 
     methods: {
-        onBlur(param) {
-            this.clearServerError(param);
-            this.v$.formData[param].$touch();
-        },
-
         close() {
             this.$emit('close');
             this.v$.formData.$reset();
@@ -64,44 +62,6 @@ export default {
 
         clearInputs() {
             this.formData = { ...this.defaultForm };
-        },
-
-        handleError(param) {
-            const { formData } = this.v$;
-
-            if (!formData[param].$error) {
-                return this.getServerError(param);
-            }
-
-            const vError = formData.$errors.find(
-                error => error.$property === param
-            );
-
-            if (vError) {
-                return vError.$message;
-            }
-
-            return 'Something is wrong there.';
-        },
-
-        getServerError(param) {
-            if (this.serverErrors.length) {
-                const serverError = this.serverErrors.find(
-                    error => error.param === param
-                );
-
-                if (serverError) {
-                    return serverError.message;
-                }
-            }
-
-            return '';
-        },
-
-        clearServerError(param) {
-            this.serverErrors = this.serverErrors.filter(
-                error => error.param !== param
-            );
         },
 
         onSuccess() {
